Remove empty lines.
Lines with only blanks in them are not written out as this makes the
files smaller and pages can run faster. It can also be used for
animation as you only update the lines that have movement. On the
negative side, this can cause problems in transmission if you don't use
the erase page flag, the old line can get left behind on screen. If you
need to force blank lines out then put an invisible code as the first
character, such as Black Background or any colour.