Packet 27 stuff
Fastext links default to 8ff. If 8ff then the FL is not written out. If
you mess with packet 27 this allows you to stop Fastext from
interfering. Row numbers are extended to allow 26 and 27. These extra
packets are read and saved but otherwise not processed by wxTED. If you
place a hand crafted packet in these rows it will be picked up by VBIT.
As these rows are difficult to cruft by hand Alistair Cree has written a
subcode link generator in
http://zxnet.co.uk/teletext/teletextlinkpacket.html